* Changed plane counter to int
* uint16_t was causing arithmetic overflows

+++ src/tools/ufo2map/bsp.h
@@ -67,7 +67,7 @@ typedef struct {
 } tree_t;
 
 extern plane_t mapplanes[MAX_MAP_PLANES];
-extern uint16_t nummapplanes;
+extern int nummapplanes;
 
 extern int nummapbrushes;
 extern mapbrush_t mapbrushes[MAX_MAP_BRUSHES];
